Refactor faulty transformers and analysis config files
