+++ src/test/java/cloud/filibuster/functional/java/redis/scenarios/JUnitRedisFilibusterFaultyHasNextFalseStringTransformerTest.java
@@ -5,7 +5,7 @@ import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.lettuce.RedisInterceptorFactory;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.FaultyHasNextFalseRedisTransformStringAnalysisConfigurationFile;
+import cloud.filibuster.functional.java.redis.transformers.configuration.FaultyHasNextFalseRedisTransformStringAnalysisConfigurationFile;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
 import org.junit.jupiter.api.BeforeAll;
