+++ src/test/java/cloud/filibuster/functional/java/redis/transformers/FaultyStringTransformerWithFalseHasNext.java
@@ -1,6 +1,8 @@
-package cloud.filibuster.junit.server.core.transformers;
+package cloud.filibuster.functional.java.redis.transformers;
 
 import cloud.filibuster.exceptions.filibuster.FilibusterFaultInjectionException;
+import cloud.filibuster.junit.server.core.transformers.Accumulator;
+import cloud.filibuster.junit.server.core.transformers.Transformer;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import com.google.gson.reflect.TypeToken;
